+++ src/stocksignals/calc_moving_avg.py
